python/pip_requirements.txt
Babel
Biggus
Cartopy
Cython
ExifRead
Iris
Jinja2
MarkupSafe
PyHamcrest
PyYAML
Pygments
Shapely
Sphinx
coards
coverage
docopt
docutils
elasticsearch
h5py
jasmin-cis
mock
netCDF4
nose
numpy
pyhdf
pyhull
pyinotify
pyke
pyshp
python-coveralls
pytz
requests
scipy
sh
simplejson
six
snowballstemmer
urllib3
wheel
xmltodict
